rename cassandra's MurmurHash
